legacy-client depends on shared.test
