+++ packages/legacy-client/package.json
@@ -37,6 +37,7 @@
     "tiny-invariant": "^1.3.1"
   },
   "devDependencies": {
+    "@osdk/shared.test": "workspace:*",
     "@types/luxon": "^3.3.3",
     "@types/ngeohash": "^0.6.4",
     "ts-expect": "^1.3.0",
