+++ packages/legacy-client/src/client/actions/actions.test.ts
@@ -16,6 +16,14 @@
 
 import { createClientContext } from "@osdk/shared.net";
 import type { ClientContext } from "@osdk/shared.net";
+import {
+  expectFetchToBeCalledWithBody,
+  expectFetchToBeCalledWithGet,
+  getMockTodoObject,
+  MOCK_ORIGIN,
+  mockFetchResponse,
+  MockOntology,
+} from "@osdk/shared.test";
 import type { MockedFunction } from "vitest";
 import {
   assert,
@@ -35,14 +43,6 @@ import {
   type Result,
   ReturnEditsMode,
 } from "../..";
-import { MockOntology } from "../../util/test";
-import {
-  expectFetchToBeCalledWithBody,
-  expectFetchToBeCalledWithGet,
-  mockFetchResponse,
-} from "../../util/test/fetchUtils";
-import { MOCK_ORIGIN } from "../../util/test/mocks/mockMetadata";
-import { getMockTodoObject } from "../../util/test/mocks/mockObjects";
 import { unwrapResultOrThrow } from "../../util/test/resultUtils";
 import { createBaseOsdkObjectSet } from "../objectSets/OsdkObjectSet";
 import type { OsdkLegacyObjectFrom } from "../OsdkLegacyObject";
