+++ packages/legacy-client/src/client/net/getLastPoint.test.ts
@@ -17,14 +17,14 @@
 import type { TimeSeriesPoint } from "@osdk/gateway/types";
 import type { ClientContext } from "@osdk/shared.net";
 import { createClientContext } from "@osdk/shared.net";
-import type { MockedFunction } from "vitest";
-import { beforeEach, describe, expect, it, vi } from "vitest";
-import { MockOntology } from "../../util/test";
 import {
   expectFetchToBeCalledWithGet,
+  MOCK_ORIGIN,
   mockFetchResponse,
-} from "../../util/test/fetchUtils";
-import { MOCK_ORIGIN } from "../../util/test/mocks/mockMetadata";
+  MockOntology,
+} from "@osdk/shared.test";
+import type { MockedFunction } from "vitest";
+import { beforeEach, describe, expect, it, vi } from "vitest";
 import { unwrapResultOrThrow } from "../../util/test/resultUtils";
 import { Timestamp } from "../baseTypes";
 import { getLastPoint } from "./getLastPoint";
