+++ packages/legacy-client/src/client/objectSets/mapPropertiesToGroupByProperties.test.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
+import { MockOntology } from "@osdk/shared.test";
 import { describe, expect, it } from "vitest";
-import { MockOntology } from "../../util/test";
 import { mapPropertiesToGroupByProperties } from ".";
 import { GroupKeyType } from "./aggregations/groupBy/GroupKeyType";
