+++ packages/legacy-client/src/util/test/ObjectTypeWithAllPropertyTypes.ts
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import type { ObjectTypeDefinition } from "@osdk/api";
 import type {
   Attachment,
   GeoPoint,
@@ -57,130 +56,3 @@ export interface ObjectTypeWithAllPropertyTypes extends OntologyObject {
   readonly geoPointArray: GeoPoint[] | undefined;
   readonly geoShapeArray: GeoShape[] | undefined;
 }
-
-export const ObjectTypeWithAllPropertyTypes = {
-  apiName: "ObjectTypeWithAllPropertyTypes",
-  primaryKeyType: "integer",
-  links: {},
-  properties: {
-    id: {
-      multiplicity: false,
-      type: "integer",
-    },
-    string: {
-      multiplicity: false,
-      type: "string",
-    },
-    boolean: {
-      multiplicity: false,
-      type: "boolean",
-    },
-    date: {
-      multiplicity: false,
-      type: "datetime",
-    },
-    dateTime: {
-      multiplicity: false,
-      type: "timestamp",
-    },
-    decimal: {
-      multiplicity: false,
-      type: "decimal",
-    },
-    integer: {
-      multiplicity: false,
-      type: "integer",
-    },
-    long: {
-      multiplicity: false,
-      type: "long",
-    },
-    short: {
-      multiplicity: false,
-      type: "short",
-    },
-    float: {
-      multiplicity: false,
-      type: "float",
-    },
-    double: {
-      multiplicity: false,
-      type: "double",
-    },
-    byte: {
-      multiplicity: false,
-      type: "byte",
-    },
-    attachment: {
-      multiplicity: false,
-      type: "attachment",
-    },
-    geoPoint: {
-      multiplicity: false,
-      type: "geopoint",
-    },
-    geoShape: {
-      multiplicity: false,
-      type: "geoshape",
-    },
-    stringArray: {
-      multiplicity: true,
-      type: "string",
-    },
-    booleanArray: {
-      multiplicity: true,
-      type: "boolean",
-    },
-    dateArray: {
-      multiplicity: true,
-      type: "datetime",
-    },
-    dateTimeArray: {
-      multiplicity: true,
-      type: "timestamp",
-    },
-    decimalArray: {
-      multiplicity: true,
-      type: "decimal",
-    },
-    integerArray: {
-      multiplicity: true,
-      type: "integer",
-    },
-    longArray: {
-      multiplicity: true,
-      type: "long",
-    },
-    shortArray: {
-      multiplicity: true,
-      type: "short",
-    },
-    floatArray: {
-      multiplicity: true,
-      type: "float",
-    },
-    doubleArray: {
-      multiplicity: true,
-      type: "double",
-    },
-    byteArray: {
-      multiplicity: true,
-      type: "byte",
-    },
-    attachmentArray: {
-      multiplicity: true,
-      type: "attachment",
-    },
-    geoPointArray: {
-      multiplicity: true,
-      type: "geopoint",
-    },
-    geoShapeArray: {
-      multiplicity: true,
-      type: "geoshape",
-    },
-  },
-} satisfies ObjectTypeDefinition<
-  "ObjectTypeWithAllPropertyTypes",
-  "ObjectTypeWithAllPropertyTypes"
->;
