+++ packages/legacy-client/tsconfig.json
@@ -18,6 +18,9 @@
     },
     {
       "path": "../shared.net"
+    },
+    {
+      "path": "../shared.test"
     }
   ]
 }
