+++ packages/shared.test/src/index.ts
@@ -15,5 +15,6 @@
  */
 
 export * as handlers from "./handlers";
+export * from "./mock-ontology";
 export { apiServer } from "./setupServers";
 export * as stubData from "./stubs";
