+++ packages/shared.test/src/mock-ontology/ObjectTypeWithReservedNames.ts
