+++ packages/shared.test/tsconfig.json
@@ -11,6 +11,9 @@
   ],
   "references": [
     {
+      "path": "../api"
+    },
+    {
       "path": "../cli"
     },
     {
